# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_api_actress.Actress:
    properties:
      aliases:
        description: Pipe-separated
        type: string
      created_at:
        type: string
      dmm_id:
        description: Real DMM actress ID when available (unique only for values >
          0)
        type: integer
      first_name:
        type: string
      id:
        type: integer
      japanese_name:
        type: string
      last_name:
        type: string
      thumb_url:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_api_actress.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/fedora-oss/javinizer-go/issues
  description: REST API for JAV metadata scraping and file organization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/fedora-oss/javinizer-go
  title: Javinizer API
  version: "1.0"
paths:
...
  /api/v1/actresses/search:
    get:
      description: Search for actresses by name (first, last, or Japanese)
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api_actress.Actress'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_actress.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_actress.ErrorResponse'
      summary: Search actresses
      tags:
      - actress
...
schemes:
- http
- https
swagger: "2.0"