# internal/api/actress/aliases.go
package actress

import (
	"github.com/fedora-oss/javinizer-go/internal/api/contracts"
	"github.com/fedora-oss/javinizer-go/internal/models"
)

type ErrorResponse = contracts.ErrorResponse
type ActressMergePreviewRequest = contracts.ActressMergePreviewRequest
type ActressMergeConflict = contracts.ActressMergeConflict
type ActressMergePreviewResponse = contracts.ActressMergePreviewResponse
type ActressMergeRequest = contracts.ActressMergeRequest
type ActressMergeResponse = contracts.ActressMergeResponse
type Actress = models.Actress