# internal/api/batch/aliases.go
package batch

import (
	"github.com/fedora-oss/javinizer-go/internal/api/contracts"
	"github.com/fedora-oss/javinizer-go/internal/api/core"
)

type ServerDependencies = core.ServerDependencies

type ErrorResponse = contracts.ErrorResponse
type BatchScrapeRequest = contracts.BatchScrapeRequest
type BatchScrapeResponse = contracts.BatchScrapeResponse
type BatchFileResult = contracts.BatchFileResult
type BatchFileResultSlim = contracts.BatchFileResultSlim
type BatchJobResponse = contracts.BatchJobResponse
type BatchJobResponseSlim = contracts.BatchJobResponseSlim
type OrganizeRequest = contracts.OrganizeRequest
type UpdateRequest = contracts.UpdateRequest
type OrganizePreviewRequest = contracts.OrganizePreviewRequest
type OrganizePreviewResponse = contracts.OrganizePreviewResponse
type UpdateMovieRequest = contracts.UpdateMovieRequest
type PosterCropRequest = contracts.PosterCropRequest
type PosterCropResponse = contracts.PosterCropResponse
type PosterFromURLRequest = contracts.PosterFromURLRequest
type PosterFromURLResponse = contracts.PosterFromURLResponse
type BatchRescrapeRequest = contracts.BatchRescrapeRequest
type BatchRescrapeResponse = contracts.BatchRescrapeResponse
type BatchExcludeRequest = contracts.BatchExcludeRequest
type BatchExcludeResponse = contracts.BatchExcludeResponse
type BatchExcludeFailed = contracts.BatchExcludeFailed
type BulkRescrapeRequest = contracts.BulkRescrapeRequest
type BulkRescrapeResponse = contracts.BulkRescrapeResponse
type BulkRescrapeMovieResult = contracts.BulkRescrapeMovieResult
type MovieResponse = contracts.MovieResponse